requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
absl-py=2.1.0=pypi_0
accimage=0.2.0=py38h3c79746_5
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
astor=0.8.1=pypi_0
asttokens=2.4.1=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
backcall=0.2.0=pypi_0
bash=0.6=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
ca-certificates=2024.7.4=hbcca054_0
cachetools=5.3.3=pypi_0
certifi=2024.7.4=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cmake=3.30.0=pypi_0
comet-ml=3.43.2=pypi_0
comm=0.2.2=pypi_0
configobj=5.0.8=pypi_0
contourpy=1.1.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.10=pypi_0
debugpy=1.8.2=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dulwich=0.22.1=pypi_0
entrypoints=0.3=pypi_0
events=0.5=pypi_0
everett=3.1.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.20.0=pypi_0
filelock=3.15.4=pypi_0
flake8=3.7.9=pypi_0
flask=2.0.3=pypi_0
flask-cors=4.0.1=pypi_0
flatbuffers=24.3.25=pypi_0
fonttools=4.53.0=pypi_0
fqdn=1.5.1=pypi_0
fsspec=2024.6.1=pypi_0
ftfy=6.2.0=pypi_0
future=1.0.0=pypi_0
gast=0.4.0=pypi_0
gdown=5.2.0=pypi_0
gevent=24.2.1=pypi_0
gevent-websocket=0.10.1=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.37=pypi_0
google-auth=2.31.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
greenlet=3.1.1=pypi_0
grpcio=1.64.1=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
haven=2.0.4=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.4=pypi_0
idna=3.7=pypi_0
imageio=2.35.1=pypi_0
importlib-metadata=8.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
iniconfig=2.0.0=pypi_0
intel-ipp=2019.1.144=h711154d_3
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
isort=4.3.21=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=pypi_0
joblib=1.4.2=pypi_0
json5=0.9.25=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.1=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.3=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
keras=2.13.1=pypi_0
kiwisolver=1.4.5=pypi_0
labelimg=1.8.6=pypi_0
labeling=0.1.14=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=14.1.0=h77fa898_0
libjpeg-turbo=2.1.5.1=hd590300_1
libstdcxx-ng=11.2.0=h1234567_1
lit=18.1.8=pypi_0
littleutils=0.2.4=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lxml=5.2.2=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mccabe=0.6.1=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpmath=1.3.0=pypi_0
mss=9.0.1=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
netkit=3.1.12=pypi_0
networkx=3.1=pypi_0
notebook=7.2.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
ogb=1.3.6=pypi_0
opencv-python=4.10.0.84=pypi_0
openssl=3.3.1=h4bc722e_2
opt-einsum=3.3.0=pypi_0
outdated=0.2.2=pypi_0
overrides=7.7.0=pypi_0
packaging=24.1=pypi_0
paddlepaddle-gpu=2.6.1=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.4.0=pypi_0
pip=24.0=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.2=pypi_0
pluggy=1.5.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.47=pypi_0
protobuf=4.25.3=pypi_0
psutil=6.0.0=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycodestyle=2.5.0=pypi_0
pycparser=2.22=pypi_0
pyflakes=2.1.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.2=pypi_0
pypdf2=3.0.1=pypi_0
pyqt5=5.12=pypi_0
pyqt5-qt5=5.15.14=pypi_0
pyqt5-sip=4.19.19=pypi_0
pysocks=1.7.1=pypi_0
pytest=7.4.4=pypi_0
python=3.8.19=h955ad1f_0
python-box=6.1.0=pypi_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=2.0.7=pypi_0
python_abi=3.8=2_cp38
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.3=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.1=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
semantic-version=2.10.0=pypi_0
send2trash=1.8.3=pypi_0
sentry-sdk=2.7.1=pypi_0
setproctitle=1.3.6=pypi_0
setuptools=69.5.1=py38h06a4308_0
simplejson=3.19.2=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
sudo=1.0.0=pypi_0
sympy=1.12.1=pypi_0
tabulate=0.9.0=pypi_0
tb-nightly=2.18.0a20240721=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.1=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.3.2=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
torch=2.0.0=pypi_0
torchaudio=2.0.1=pypi_0
torchvision=0.15.1=pypi_0
tornado=6.4.1=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.42.3=pypi_0
triton=2.0.0=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2024.1=pypi_0
ultralytics=8.2.42=pypi_0
ultralytics-thop=2.0.0=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.6.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.43.0=py38h06a4308_0
widgetsnbextension=4.0.11=pypi_0
wilds=1.2.2=pypi_0
wrapt=1.16.0=pypi_0
wurlitzer=3.1.1=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yapf=0.29.0=pypi_0
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
zope-event=5.0=pypi_0
zope-interface=7.2=pypi_0
